copy over all of Aris code
