+++ autoarray/dataset/interferometer/dataset.py
@@ -13,7 +13,8 @@ from autoarray.structures.visibilities import Visibilities
 from autoarray.structures.visibilities import VisibilitiesNoiseMap
 
 from autoarray.structures.arrays import array_2d_util
-from autoarray.inversion.inversion import inversion_util
+
+from autoarray.inversion.inversion.interferometer import inversion_interferometer_util
 
 logger = logging.getLogger(__name__)
 
@@ -154,7 +155,7 @@ class Interferometer(AbstractDataset):
         logger.info("INTERFEROMETER - Computing W-Tilde... May take a moment.")
 
         curvature_preload = (
-            inversion_util.w_tilde_curvature_preload_interferometer_from(
+            inversion_interferometer_util.w_tilde_curvature_preload_interferometer_from(
                 noise_map_real=np.array(self.noise_map.real),
                 uv_wavelengths=np.array(self.uv_wavelengths),
                 shape_masked_pixels_2d=np.array(
@@ -166,7 +167,7 @@ class Interferometer(AbstractDataset):
             )
         )
 
-        w_matrix = inversion_util.w_tilde_via_preload_from(
+        w_matrix = inversion_interferometer_util.w_tilde_via_preload_from(
             w_tilde_preload=curvature_preload,
             native_index_for_slim_index=self.real_space_mask.derive_indexes.native_for_slim,
         )
