+++ autoarray/inversion/inversion/interferometer/w_tilde.py
@@ -16,6 +16,7 @@ from autoarray.preloads import Preloads
 from autoarray.structures.visibilities import Visibilities
 
 from autoarray.inversion.inversion import inversion_util
+from autoarray.inversion.inversion.interferometer import inversion_interferometer_util
 
 from autoarray.numba_util import profile_func
 
@@ -128,7 +129,7 @@ class InversionInterferometerWTilde(AbstractInversionInterferometer):
         mapper = self.cls_list_from(cls=AbstractMapper)[0]
 
         if not self.settings.use_source_loop:
-            return inversion_util.curvature_matrix_via_w_tilde_curvature_preload_interferometer_from(
+            return inversion_interferometer_util.curvature_matrix_via_w_tilde_curvature_preload_interferometer_from(
                 curvature_preload=self.w_tilde.curvature_preload,
                 pix_indexes_for_sub_slim_index=mapper.pix_indexes_for_sub_slim_index,
                 pix_size_for_sub_slim_index=mapper.pix_sizes_for_sub_slim_index,
@@ -143,7 +144,7 @@ class InversionInterferometerWTilde(AbstractInversionInterferometer):
             sub_slim_weights_for_pix_index,
         ) = mapper.sub_slim_indexes_for_pix_index_arr
 
-        return inversion_util.curvature_matrix_via_w_tilde_curvature_preload_interferometer_from_2(
+        return inversion_interferometer_util.curvature_matrix_via_w_tilde_curvature_preload_interferometer_from_2(
             curvature_preload=self.w_tilde.curvature_preload,
             native_index_for_slim_index=self.transformer.real_space_mask.derive_indexes.native_for_slim,
             pix_pixels=self.linear_obj_list[0].params,
